+++ src/aurora/herffile.cpp
@@ -148,7 +148,7 @@ const Archive::ResourceList &HERFFile::getResources() const {
 
 const HERFFile::IResource &HERFFile::getIResource(uint32 index) const {
 	if (index >= _iResources.size())
-		throw Common::Exception("Resource index out of range (%d/%d)", index, _iResources.size());
+		throw Common::Exception("Resource index out of range (%u/%u)", index, (uint)_iResources.size());
 
 	return _iResources[index];
 }
