+++ src/aurora/ndsrom.cpp
@@ -163,7 +163,7 @@ const Archive::ResourceList &NDSFile::getResources() const {
 
 const NDSFile::IResource &NDSFile::getIResource(uint32 index) const {
 	if (index >= _iResources.size())
-		throw Common::Exception("Resource index out of range (%d/%d)", index, _iResources.size());
+		throw Common::Exception("Resource index out of range (%u/%u)", index, (uint)_iResources.size());
 
 	return _iResources[index];
 }
