+++ src/aurora/rimfile.cpp
@@ -102,7 +102,7 @@ const Archive::ResourceList &RIMFile::getResources() const {
 
 const RIMFile::IResource &RIMFile::getIResource(uint32 index) const {
 	if (index >= _iResources.size())
-		throw Common::Exception("Resource index out of range (%d/%d)", index, _iResources.size());
+		throw Common::Exception("Resource index out of range (%u/%u)", index, (uint)_iResources.size());
 
 	return _iResources[index];
 }
