+++ src/engines/aurora/console.cpp
@@ -1062,7 +1062,7 @@ bool Console::printHints(const Common::UString &command) {
 
 	if (lines >= (kConsoleLines - 3)) {
 		if (!_printedCompleteWarning)
-			printf("%d completion candidates", count);
+			printf("%u completion candidates", (uint)count);
 
 		_printedCompleteWarning = true;
 
@@ -1450,7 +1450,8 @@ void Console::printList(const std::list<Common::UString> &list, size_t maxSize) 
 
 	// Print a message when we cut items
 	if (linesCut > 0) {
-		Common::UString cutMsg = Common::UString::format("(%d items cut due to history overflow)", linesCut);
+		Common::UString cutMsg =
+			Common::UString::format("(%u items cut due to history overflow)", (uint)linesCut);
 
 		print(cutMsg);
 	}
