+++ src/engines/nwn/gui/ingame/quickbar.cpp
@@ -64,7 +64,7 @@ QuickbarButton::QuickbarButton(::Engines::GUI &gui, size_t n) : NWNWidget(gui, "
 	if (invisible)
 		invisible->setInvisible(true);
 
-	NWNWidget::setTag(Common::UString::format("Quickbar%d", _buttonNumber));
+	NWNWidget::setTag(Common::UString::format("Quickbar%u", (uint)_buttonNumber));
 	_model->setTag(NWNWidget::getTag());
 
 }
