+++ src/graphics/images/cbgt.cpp
@@ -194,7 +194,7 @@ void CBGT::readCells(ReadContext &ctx) {
 			ctx.cells.back() = Aurora::Small::decompress(cellData);
 
 			if (ctx.cells.back()->size() != 4096)
-				throw Common::Exception("Invalid size for cell %u: %u", i, (uint)ctx.cells.back()->size());
+				throw Common::Exception("Invalid size for cell %u: %u", (uint)i, (uint)ctx.cells.back()->size());
 
 			ctx.cbgt->seek(pos);
 		}
